Reorganising Cogent.C so that Cogent.C.Compile can import Cogent.DataLayout.CodeGen
